refactor(catalog-dao): catalog_dao 三模块正交分解 + Façade 多继承聚合;
按职责将 510 行单文件拆为三模块：
- catalog_dao.py（Façade，169 行）：保留 Catalog 顶层 CRUD + CatalogDao 类多继承
  聚合 _CatalogTopLevelDao + CatalogNodeDao + CatalogAssignmentDao，重导出
  _NODE_TYPE_TO_ENUM / _ENUM_TO_NODE_TYPE / MAX_TREE_DEPTH / _compute_slug 兼容历史
  导入。
- catalog_node_dao.py（新，261 行）：节点条目 CRUD + 树查询（CTE 复用至
  _query_tree 共享实现 get_tree / get_subtree）。
- catalog_assignment_dao.py（新，145 行）：DOCUMENT_REF 软引用 N:M 维护
  （assign_document / unassign_document / get_node_documents / get_document_nodes）。

外部调用面零改动：CatalogDao.create_node / CatalogDao.assign_document 等仍可用；
新代码可按职责直接 import 细分类提高可读性。

测试：新增 test_catalog_dao_facade.py 锁定 Façade 多继承契约（13+ 方法可用 +
重导出符号兼容）。test_catalog_dao_unit.py 28 例零改动通过。

详见 docs/issue.md ISSUE-023 第 (2) 步。
